GEODE-2089 entry-idle-time on client side cache is not working as expected
When we pull an entry in from another cache in LocalRegion.findObjectInSystem()
we end up invoking updateStatsForPut and then updateStatsForGet.  The former
method installs a lastModifiedTime from the version tag that came from the
other cache and this is used in updateStatsForPut to schedule an expiry-task
for the entry.  Then updateStatsForGet is invoked to set the lastAccessed
time of the entry to the current time.  However, by the time this is done
the entry may have already been removed by the expiry-task if the
lastModified time was too far in the past.

The fix is to establish both the lastModified and lastAccessed times in
updateStatsForPut.

I've included two of the "leaf" RegionEntry classes in the diff but all of
them had to be modified in the same way.