GEM-1072, GEODE-2136: Don't duplicate cookies in the http response
We had some code that copied cookies from the request to the response.
That caused us to include a potentially stale cookie value in the
response.

Adding a unit test that we don't screw up the users cookies. I had to
bring in a dependency on httpunit, because the HttpTester with jetty is
not correctly parsing multiple Set-Cookie headers.

(cherry-picked from 03715a63eee5f20453f0dc0ec01311b11d7548af)