GEODE-2186: Changing the gateway status to only looking running status
Gateway senders may or may not be connected, depending on what is going
on with the underlying queue. If a sender is a secondary, it may be
disconnected because it doesn't need to send anything. Even primaries
may be disconnected for periods of time if the connection is lost when
there is nothing in the queue.

(cherry picked from commit 09184881aec9250fdf2e96f4fced4dedaac0bda7)