GEODE-1272 Don't deserialize PDX objects when creating an index
Setting the flag to prevent deserialization of PDX objects while
populating an index that is defined on a partitioned region. We were
setting this flag in the member that initially created the index, but
not in other members that receive the IndexCreationMessage.

(cherry-picked from commit 341a359e00a9ba2f9c74e1dc4fee38fc53f78b3c)