GEODE-2155 Auto-reconnect fails with NPE
I created a test to reproduce the problem by introducing a non-Declarable
cache listener and then initiating a forced-disconnect.  cache.xml is
generated and when it's used to rebuild the cache a CacheXmlException
is thrown, reliably reproducing the problem.

The fix is to simply catch the CacheXmlException in
InternalDistributedSystem.reconnect() and terminate reconnection attempt

(cherry picked from commit 638b05840d388f7d6476bad7ea0729510b67aca3)