GEODE-3043 surprise member added when the member is already in the cluster
Modified InternalDistributedMember to disregard the "name" field in compareTo
if partial IDs are being used in the comparison.

Modified GMSMembershipManager to attempt to replace partial IDs with the
full IDs from the membership view.  Typically these IDs will cause no
problems and methods in GMSMembershipManager that send messages to a
recipient will also attempt to replace the ID with a full ID from the
membership view, but it's good to try to keep IDs canonical.

(cherry picked from commit c585245cdc46393897a54602c009219c129d9daf)