GEODE-5278 Unexpected CommitConflictException caused by faulty region synchronization
Modified RegionVersionVector.contains(id,version) to answer true
if the vector is for sychronization and the ID is that of the owner
of the vector.  This causes the get-initial-image code to ignore entries
last changed by the server requesting the synch operation.

(cherry picked from commit 024ed43a0429f2df85b6477472327c3db054eb6d)